hack: Do not compile datapath
The OVS eBPF patch sets are developed against master 'net-next' tree,
which is currently targeting for kernel version 3.19.

The datapath in OVS mater does not currently support 3.19 kernel.
But we should not really using OVS datapath, The changes in the
follow patches are only changes in user space programs, which should
work with the kernel module built from the 'net-next' tree.

This patch disables building of the datapath and disables the
configuration time kernel version check.

Signed-off-by: Andy Zhou <azhou@nicira.com>