ovsdb: enhance file interface to deal with schemas
When storing joined schemas into a database file, A JSON array will
be store as its first record, with every individual schema as one
element of the array. To maintain backward compatibility, when
joined schemas contain a single schema, a single JSON object will be
stored, not a JSON array with a single element.

The joined schema is not stored into the DB file, since it can be
rebuilt at run time by rejoining stored schemas mentioned above.

DB file read operations now aceept both multiple schemas
(a JSON array as the first entry) or a single schmea (a JSON object).

This change is fully backward compatible as long as each DB contains
a single schema.

Signed-off-by: Andy Zhou <azhou@nicira.com>