ovsdb-tool: add multiple schemas support
Allow ovsdb-tool to accept a schema-list, a list of schema files, in
addition to a single schema file, for all applicable options.

There is no limit on how many schemas can be joined together, but
they can not have a shared schema name and they have to be compatible
for joining. Compatible basically means overlapping tables and columns
have to be of the same type.

A single joined schema, formed by joining schemas from the schema-list,
is used for DB operations.

See man page changes for more details.

Signed-off-by: Andy Zhou <azhou@nicira.com>