ovsdb: Generate the joined the schema name
This patch implements a simple algorithm for picking the joined
schema name. The joined schema name takes the name of one of schema
joined. Any schema name starts with 'O' will be favored over other
ones. If multiple names can be picked. Ties are broken at random.

The goal of this algorithm is to favor the default schema names, such
as OVS* or OVN*.

Signed-off-by: Andy Zhou <azhou@nicira.com>