ovsdb: add idl class call backs
When creating ovsdb IDL, allow application to register call
back functions that will can create and destroy IDL classes
based on schemas received from OVSDB server dynamically.

The dynamically created IDL class will be used for ovsdb_idl_run(),
replacing the default IDL class created the IDL compiler.
It is required that  the new schema to be an supper set of
original schema used by IDL compiler so that the IDL compiler generated
library will continue to be functional. This requirement is met when
the new schema is a product of OVSDB schema join operation.

Although the implementation is generic in nature, the intended
user is ovs-vsctl, for it to support basic OVSDB operations
against OVSDB server with joint schemas.

Signed-off-by: Andy Zhou <azhou@nicira.com>