odp-execute: Support dropping packets.
Meter action can drop or modify packets, so the execution framework is
changed to allow for this.  Also, a meter action can appear alone
(e.g., to measure traffic that is to be dropped), so the existing drop
implementation is not sufficient any more.

The action execution framework is changed in three ways:

1. Action execution can shrink the number of packets in a batch to be
   further processed by the remaining actions.

2. Whenever a packet is 'stolen' (e.g., for output), the corresponding
   packet pointer is changed to NULL. NULLed pointers must be excluded
   from further processing by using the change #1 above.  Sometimes
   this means that the packet pointers are shuffled so that remaining
   action processing never needs to check for the NULL pointers
   explicitly.

3. Packet deletion is centralized to the odp_execute_actions()
   function.  This make memory leaks less likely to appear as new
   kinds of action are added later.

Signed-off-by: Jarno Rajahalme <jarno@ovn.org>