ovsdb: Do not group sessions by remote.
Currently ovsdb_jsonrpc_session are grouped together in a linked
list within  'ovsdb_jsonrpc_remote'. This makes sense since most
session operations applies to sessions within a remote.

However, in order to scale up ovsdb-server with multi-threading, it is
more convenient to distribute a sessions to any thread available,
regardless which remote it is associated with.

This patch introduces a set of APIs that provide operations on
a list of sessions. Instead of group sessions by remote, they
are linked together in a new ovs_list field 'all_sessions' in the
ovsdb_jsonrpc_server struct.

With multi-threading, the design is that all sessions managed
by a thread will have them linked together on a thread private
linked list. At that time, the 'all_sessions' field in
ovsdb_jsonrpc_server struct will have all session managed
the main process.

Signed-off-by: Andy Zhou <azhou@ovn.org>