ovsdb: Introduce sessions API
Currently ovsdb_jsonrpc_session are grouped together in a linked
list within  'ovsdb_jsonrpc_remote'. This makes sense since most
session operations applies to sessions within a remote.

Considering scale up ovsdb-server with multi-threading, it can be
more convenient to distribute a sessions to any thread available.

This patch introduces a set of APIs that provide operations on
a list of sessions, in preparation for multi-threading to be
introduced in later patches.

Signed-off-by: Andy Zhou <azhou@ovn.org>