ovsdb-server: More threads, please.
This patch adds some basic infrastructure of creating and destroying
threads that will be used to handle jsonrpc sessions.

This patch implements a simple logic to create thread on demand.
Using number of sessions a server services as an rough measure of
system load, A jsonrpc server creates more threads as the number of
sessions exceeds a pre defined number of session per thread, Until
a user specified maximum is reached.

The threading model is simple. OVSDB server starts up with no threads
(This is required to handle the --run option).  Once created,
a thread is never destroyed until the OVSDB server shuts down.

Session thread currently does nothing useful. In fact, the threads
won't even be created because the max number threads is hard coded
to zero. Later patch will create and make use of those threads.

Signed-off-by: Andy Zhou <azhou@ovn.org>