ovsdb: Make 'remote' opaque in ovsdb_jsonrpc_session
It turns out there is no need for ovsdb_jsonrcp_session to have access
to the remote data structure. Make it opaque as a 'void *' pointer.
The pointer value is still useful when selecting ovsdb_jsonrpc_sessions
that of the same 'remote'.  This change will laster multi-threading
patches easier because access to 'remote' does not need be protected.

Signed-off-by: Andy Zhou <azhou@ovn.org>