ovsdb: Make 'remote' opaque in ovsdb_jsonrpc_session
It turns out there is no need for ovsdb_jsonrpc_session to have access
to the remote data structure. Make it opaque by introduce the definition
of ovsdb_jsonrpc_session later in the file, only above the functions
that need to operate on this data structure.

This change will later multi-threading locking easier because
access to 'remote' are only done from the main process, Each thread
only use the pointer value for matching 'remote'. There is no need
for lock protecting 'remote'.

Signed-off-by: Andy Zhou <azhou@ovn.org>