ovsdb: Add IPC messages for thread communication
Add a command queue for the main thread to pass commands to the
sessions thread.

Defined an implemented a set of IPC messages for creating,
closing and managing jsonrpc sessions from the main process to jsonrpc
sessions threads.

Add a doorbell mechanism for a jsonrpc session to inform other sessions
of possible OVSDB changes.

Although all supporting functions for running a multithreading OVSDB
server have been implemented, multithread is still not turned on,
because 'n_max_threads' is still hard coded to zero.  Future patches
will remove this restriction.

Signed-off-by: Andy Zhou <azhou@ovn.org>