ovsdb: Replication usability improvements
Added the '--no-sync' option base on feedbacks of current
implementation.

Added appctl command "ovsdb-server/sync-status" based on feedbacks
of current implementation.

Added a test to simulate the integration of HA manager with OVSDB
server using replication.

Other documentation and API improvements.

Signed-off-by: Andy Zhou <azhou@ovn.org>
------

I hope to get some review comments on the command line and appctl
interfaces for replication. Since 2.6 is the first release of those
interfaces, it is easier to making changes, compare to future
releases.